fix(sync): resolve 7 sync push contract bugs in Batch C remediation
1. Workout mode encoding: workoutModeToSync now handles DB-stored class
   names (OldSchool, TUTBeast) via programModesByClassName map, plus
   display names and sync strings. Prevents OLDSCHOOL instead of OLD_SCHOOL.

2. Telemetry setId mismatch: moved telemetry building into
   buildPortalExerciseWithTelemetry where the generated setId is in scope,
   ensuring FK consistency. Added ExerciseWithTelemetry bundle and
   PortalSessionBuildResult to carry both exercises and telemetry.

3. VBT assessment duplication: use stable result.id.toString() instead
   of generateUUID() in toPortalAssessmentResult, preventing duplicate
   rows on every sync push.

4. prCount hardcoded to zero: replaced val totalPrs = 0 with
   sorted.count { it.isPr } to compute actual PR count from session data.

5. Mutex concurrency guard: added syncMutex.withLock to sync() to
   prevent concurrent sync races.

6. Stale lastSync for pull: pullRemoteChanges now receives syncTimeEpoch
   from the push response instead of the stale tokenStorage timestamp.

7. totalVolumeKg Float precision: changed GamificationStatsSyncDto field
   from Int to Float, removed .toInt() truncation in PortalPullAdapter.

+++ shared/src/commonMain/kotlin/com/devil/phoenixproject/data/sync/PortalPullAdapter.kt
@@ -65,7 +65,7 @@ object PortalPullAdapter {
             clientId = "gamification_stats_1", // Singleton row
             totalWorkouts = stats.totalWorkouts,
             totalReps = stats.totalReps,
-            totalVolumeKg = stats.totalVolumeKg.toInt(), // Legacy DTO uses Int
+            totalVolumeKg = stats.totalVolumeKg,
             longestStreak = stats.longestStreak,
             currentStreak = stats.currentStreak,
             updatedAt = now
